+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Adjust map header output type with lookup id
+
+
 ## 7.7.1 Feb 6, 2022
 
 Upgrade priority: Low. Recommended for users with long-running scripts.